refactor: move the document history to the store
