+++ src/services/ghost/GhostContext.ts
@@ -10,58 +10,14 @@ export class GhostContext {
 		this.documentStore = documentStore
 	}
 
-	/**
-	 * Get the last 10 operations performed by the user on a document
-	 * @param document The document to get operations for
-	 * @returns A diff string representing the last 10 operations
-	 */
-	private getRecentOperations(document: vscode.TextDocument): string {
-		if (!document) {
-			return ""
-		}
-
-		const uri = document.uri.toString()
-		const item = this.documentStore.getDocument(document.uri)
-
-		if (!item || item.history.length < 2) {
-			return ""
-		}
-
-		// Get the last 10 versions (or fewer if not available)
-		const historyLimit = 10
-		const startIdx = Math.max(0, item.history.length - historyLimit)
-		const recentHistory = item.history.slice(startIdx)
-
-		// If we have at least 2 versions, compare the oldest with the newest
-		if (recentHistory.length >= 2) {
-			const oldContent = recentHistory[0]
-			const newContent = recentHistory[recentHistory.length - 1]
-
-			// Generate a diff between the oldest and newest versions
-			const filePath = vscode.workspace.asRelativePath(document.uri)
-			const diffPatch = createPatch(filePath, oldContent, newContent, "Previous version", "Current version")
-
-			return diffPatch
-		}
-
-		return ""
-	}
-
-	/**
-	 * Add recent user operations to the context
-	 * @param context The context to add operations to
-	 * @returns The updated context
-	 */
 	private addRecentOperations(context: GhostSuggestionContext): GhostSuggestionContext {
 		if (!context.document) {
 			return context
 		}
-
-		const recentOperations = this.getRecentOperations(context.document)
+		const recentOperations = this.documentStore.getRecentOperations(context.document)
 		if (recentOperations) {
 			context.recentOperations = recentOperations
 		}
-
 		return context
 	}
 
@@ -90,7 +46,6 @@ export class GhostContext {
 		if (!context.document) {
 			return context
 		}
-
 		if (this.documentStore.needsASTUpdate(context.document)) {
 			await this.documentStore.storeDocument(context.document, true)
 		}
